pyproject.toml
[project]
name = "auto-feedback-backend"
version = "0.1.0"
description = "The Backend of Auto Feedback Generator for CourseVita"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask==3.1.1",
    "flask-sqlalchemy==3.1.1",
    "flask-jwt-extended==4.7.1",
    "python-dotenv==1.1.1",
    "openai==1.93.0",
    "werkzeug==3.1.3",
]

[tool.uv]
dev-dependencies = []
